pyproject.toml
[tool.poetry]
name = "morse-api"
version = "0.1.0"
description = ""
authors = ["SherlockH0 <ddmlf7@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9.0"
django = "4.2.9"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"
django-cors-headers = "^4.4.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "^2.9.0"
django-cryptography = "^1.1"
channels = "^4.1.0"
daphne = "^4.1.2"
cloudinary = "^1.41.0"


[tool.poetry.group.dev.dependencies]
django-types = "^0.19.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
